Add user controller
