+++ config.ru
@@ -4,6 +4,7 @@ require File.join(File.dirname(__FILE__), "api", "app", "controllers", "base_app
 require File.join(File.dirname(__FILE__), "api", "app", "controllers", "api")
 require File.join(File.dirname(__FILE__), "api", "app", "controllers", "auth")
 require File.join(File.dirname(__FILE__), "api", "app", "controllers", "data_app")
+require File.join(File.dirname(__FILE__), "api", "app", "controllers", "user_app")
 
 $stdout.sync = true
 
@@ -21,4 +22,8 @@ end
 
 map "/api/data" do
   run DataApp.new
+end
+
+map "/api/user" do
+  run UserApp.new
 end
